+++ packages/client/hmi-client/src/components/widgets/tera-grey-card.vue
